Added local test for GetCustomerList
